requirements.txt
python-dotenv>=1.0.1
python-telegram-bot>=21.10,<22
langchain>=0.3.20
langchain-community>=0.3.19
langchain-core>=0.3.41
langchain-text-splitters>=0.3.6
langchain-ollama>=0.2.3
langchain-chroma>=0.2.2
chromadb>=0.6.3
pymupdf>=1.25.3
pytesseract>=0.3.13
pillow>=11.1.0
pypdf>=5.3.0
